Add style improvements and refactor DomAndEvents.js
Implemented various improvements in CSS for better readability and usability. Padding and resizing of table headers and columns were adjusted along with the deletion of unnecessary code for a cleaner style. This also included deletion of duplicated CSS rules and normalization of font styles.

The Javascript file 'DomAndEvents.js' was refactored to improve error handling and the overall clarity of the code. Changes include more specific targeting for update and delete buttons, splitting up common functionality into separate methods, introduction of 'deletePopup' for delete confirmation, and better error handling in 'Crud.js'. This has increased the maintainability and readability of the code.

These updates were made to improve the user experience and to make the codebase easier to manage and understand.